pyproject.toml
[project]
name = "local-project-manager"
version = "0.1.0"
description = "TUI for managing local development projects"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.45",
    "rich>=14.2.0",
    "textual>=6.4.0",
    "tomli>=2.3.0",
]

[project.scripts]
lpm = "lpm.app:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src/lpm",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["src/lpm"]
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
